deadlock detection: add interface for dumping runtime execution graph
Added include/seastar/core/internal/deadlock_utils.hh and src/core/deadlock_utils.cc.

They both are enabled if SEASTAR_DEADLOCK_DETECTION flag is set.
They contain function for tracing evolution of runtime graph of tasks, promises and futures with info about semaphores.
Runtime vertex is task, future_base or promise_base and can be converted implicitly.
Example edges are:
- from promise to it's continuation if then was called before promise being ready,
- from promise to future and from future to continuation if then was called on ready future,
- from continuation to promise that represented finishing of continuation.
There is also tracing of speculative edges (we don't know the edge from current vertex to first in continuation chain, because it can be created anywhere.
Current traced vertex ptr can be updated using updated_current_traced_vertex struct. It works with RAII semantics updating for its lifetime and reverts when destructed.
Also there are functions to trace semaphore events (waits, signals) and corresponding runtime vertex.

The struct runtime_vertex is used to have type saved in enum to allow for passing it to deadlock_utils.cc file in order to avoid cyclic dependencies.
As for semaphores, they are just cast to void * (which is not possible with futures, etc, because address of future_base can be different that address of future).
Also when detection is disabled there are empty stubs for each function to allow calling them not under #ifdef.

Tracing code is a simple std::ostream with JSON serialization of available data for now.

Co-authored-by: Antoni Żewierżejew <azewierzejew@gmail.com>